@@ -17,16 +17,13 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@hookform/resolvers": "^2.8.3",
     "@tanstack/react-table": "^8.10.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
     "@types/js-cookie": "^3.0.4",
     "@types/leaflet": "^1.7.5",
     "@types/lodash": "^4.14.176",
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-relay": "^14.1.5",
-    "@types/relay-runtime": "^14.1.13",
-    "@types/relay-test-utils": "^14.1.0",
+    "@types/react": "^18.2.25",
+    "@types/react-dom": "^18.2.10",
+    "@types/react-relay": "^14.1.6",
+    "@types/relay-runtime": "^14.1.14",
     "@types/semver": "^7.3.13",
     "@vitejs/plugin-react": "^4.0.4",
     "apexcharts": "^3.42.0",
@@ -35,20 +32,19 @@
     "graphql": "^15.7.2",
     "history": "^5.1.0",
     "js-cookie": "^3.0.5",
-    "leaflet": "^1.7.1",
+    "leaflet": "^1.9.4",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-apexcharts": "^1.4.1",
     "react-bootstrap": "^2.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^3.1.4",
     "react-hook-form": "^7.46.2",
     "react-intl": "^6.4.7",
-    "react-leaflet": "^3.2.2",
+    "react-leaflet": "^4.2.1",
     "react-relay": "^15.0.0",
-    "react-router-dom": "^6.0.0-beta.8",
+    "react-router-dom": "^6.16.0",
     "react-select": "^5.7.5",
-    "react-select-event": "^5.5.0",
     "relay-compiler": "^15.0.0",
     "relay-runtime": "^15.0.0",
     "relay-test-utils": "^15.0.0",
@@ -99,7 +95,10 @@
     "react-error-overlay": "6.0.9"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.1.2",
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
+    "@types/relay-test-utils": "^14.1.1",
     "@typescript-eslint/eslint-plugin": "^6.7.2",
     "@typescript-eslint/parser": "^6.7.2",
     "@vitest/coverage-v8": "^0.34.5",
@@ -110,6 +109,7 @@
     "eslint-plugin-relay": "^1.8.3",
     "jsdom": "^22.1.0",
     "prettier": "^3.0.3",
+    "react-select-event": "^5.5.1",
     "vitest": "^0.34.5"
   }
 }
